* Removing symbolic links in experiments/data/finance (closes #275) * Adding PostgreSQL result files for all TPCH and financial queries. * Code formatting of Scala source files (closes #202). * Resolving problems with DB parser. All TPCH result files use '|' as delimiter and quotes around strings. * Minor tweaks to query_test.rb: the temporary file is deleted at exit, db_parser is extended to support semicolons in strings.
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@2554 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 1ad0f8e8836a249d44ecc922a201e5515074f604

+++ test/scripts/postgres/tpch/run_query.sh
@@ -2,25 +2,50 @@
 
 if [ $# -lt 2 ]
 then
-  echo "Usage: `basename $0` (query3|query11|...) (standard|big|tiny|standard_del|...) [delimiter=',']"
+  echo "Usage: `basename $0` (all|query3|query11|...) (all|standard|big|tiny|standard_del|...) [delimiter='|']"
   exit 1
 fi
 
 SCRIPT_DIR=`dirname $0`
+COMMAND=$0
 QUERY=$1
 DATASET=$2
 OUTPUT_FILE=${QUERY}_${DATASET}.csv
-DELIMITER=","
+DELIMITER="|"
 
 if [ $# -gt 2 ]
 then
   DELIMITER=$3
 fi
 
-echo "Running $QUERY on $DATASET"
+if [ $QUERY = "all" ]
+then
+ 
+  QUERIES=`ls queries/*.sql`
+  for QUERY_EXT in `ls queries/*.sql` 
+  do
+    QUERY=`basename $QUERY_EXT .sql`
+    $COMMAND $QUERY $DATASET $DELIMITER
+  done
 
-sed -e "s:@@DATASET@@:$DATASET:g" $SCRIPT_DIR/queries/$QUERY.sql > _query.sql
-psql -AqtF $DELIMITER -f _query.sql -o $OUTPUT_FILE
-rm -f _query.sql
+else 
+  if [ $DATASET = "all" ]
+  then
 
-echo "Done. The result is in $OUTPUT_FILE."
+    for DATASET in tiny tiny_del standard standard_del big big_del 
+    do
+      $COMMAND $QUERY $DATASET $DELIMITER
+    done
+
+  else
+
+    echo "Running $QUERY on $DATASET"
+
+    sed -e "s:@@DATASET@@:$DATASET:g" $SCRIPT_DIR/queries/$QUERY.sql > _query.sql
+    psql -AqtF $DELIMITER -f _query.sql -o $OUTPUT_FILE
+    rm -f _query.sql
+
+    echo "Done. The result is in $OUTPUT_FILE."
+
+  fi
+fi
